Remove most callers of repository_for_path
